+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * [#5801](https://github.com/bbatsov/rubocop/pull/5801): Add new `Rails/RefuteMethods` cop. ([@koic][])
+* [#5805](https://github.com/bbatsov/rubocop/pull/5805): Add new `Rails/AssertNot` cop. ([@composerinteralia][])
 
 ### Bug fixes
 
@@ -3328,3 +3329,4 @@
 [@Darhazer]: https://github.com/Darhazer
 [@Wei-LiangChew]: https://github.com/Wei-LiangChew
 [@svendittmer]: https://github.com/svendittmer
+[@composerinteralia]: https://github.com/composerinteralia